Connection & Schema moved from server.js, cleaning
