+++ controllers/router.js
@@ -56,3 +56,4 @@ router.put('/:id', (req, res) => {
     })
 })
 
+module.exports = router;
