+++ server.js
@@ -1,29 +1,24 @@
+////// Importing Dependencies //////
 require("dotenv").config()
-const { PORT = 3000, DATABASE_URL } = process.env
-// or const PORT = process.env.PORT || 3000
+const { PORT = 3000 } = process.env // or const PORT = process.env.PORT || 3000
 const express = require("express")
 const cors = require("cors")
 const morgan = require("morgan")
+const bookmarkRouter = require('./controllers/router')
+
+////// Application Object //////
 const app = express()
-const mongoose = require("mongoose")
-mongoose.connect(DATABASE_URL, {
-    useUnifiedTopology: true,
-    useNewUrlParser: true
-})
-mongoose.connection
-.on("open", () => console.log("Mongoose connected"))
-.on("close", () => console.log("Disconnected from Mongoose"))
-.on("error", (error) => console.log(error))
-const BookmarksSchema = new mongoose.Schema({
-    title: String,
-    url: String
-})
-const Bookmarks = mongoose.model("Bookmarks", BookmarksSchema)
-app.use(cors())
-app.use(morgan("dev"))
-app.use(express.json())
+
+////// Middleware //////
+app.use(cors()) // Allows requests from external sites
+app.use(morgan("dev")) // Logs requests to our url
+app.use(express.json()) // Parses incoming json
+
 ////// Routes //////
 app.get("/", (req, res) =>{
-    res.send("Let's do this team 1")
+    res.send("Let's do this team !")
 })
+
+////// App Listener //////
 app.listen(PORT, () => console.log(`Server is listening on PORT ${PORT}`))
+
